test: split app-view tests into message and conversation mode
