+++ src/views/tests/app-view-conversation-mode.test.tsx
@@ -17,12 +17,10 @@ import {
 import { conversationTestUtilities } from '../../__test__/conversation/ui-interactions';
 import {
 	generateConversationFromAPI,
-	generateConvMessageFromAPI,
-	generateMessageFromAPI
+	generateConvMessageFromAPI
 } from '../../__test__/generators/api';
 import { mockLayoutStorage } from '../../__test__/layouts-utils';
-import { stubSearchMessages } from '../../__test__/message/api-stub';
-import { setupViewByConversation, setupViewByMessage } from '../../__test__/setup-utils';
+import { setupViewByConversation } from '../../__test__/setup-utils';
 import { MAILS_VIEW_LAYOUTS, MAILS_VIEW_SPLIT_LAYOUT_ORIENTATIONS } from '../../constants';
 import AppView from '../app-view';
 import { makeAllItemsVisible } from '../settings/filters/tests/test-utils';
@@ -30,7 +28,7 @@ import { setupTest } from '@test-setup';
 import { createSoapAPIInterceptor } from '@test-utils/network/msw/create-api-interceptor';
 import { populateFoldersStore } from '@test-utils/store/folders';
 
-describe('AppView', () => {
+describe('AppView in conversation mode', () => {
 	beforeAll(() => {
 		configure({ asyncUtilTimeout: 5000 });
 	});
@@ -41,27 +39,6 @@ describe('AppView', () => {
 		});
 		populateFoldersStore();
 	});
-	describe('Messages', () => {
-		it('should display received messages on app load', async () => {
-			setupViewByMessage();
-			const incompleteMessage = generateMessageFromAPI({
-				id: '123',
-				su: 'Test message 1',
-				l: '2',
-				fr: 'Test m'
-			});
-			stubSearchMessages({ messages: [incompleteMessage] });
-			setupTest(<AppView />, {
-				initialEntries: [`/folder/2`]
-			});
-
-			// lazy components need longer timeout
-			await screen.findByTestId('message-item-123', {}, { timeout: 10000 });
-			makeAllItemsVisible();
-			expect(await screen.findByText('Test message 1')).toBeInTheDocument();
-		});
-	});
-
 	describe('Conversations', () => {
 		beforeEach(() => {
 			setupViewByConversation();
